rename hazen.py to main.py
