+++ hazenlib/__main__.py
@@ -98,7 +98,7 @@ import os
 from docopt import docopt
 import pydicom
 from hazenlib.logger import logger
-from hazenlib.utils import is_dicom_file, get_dicom_files
+from hazenlib import is_dicom_file, get_dicom_files
 from hazenlib._version import __version__
